backend/ai_analyze_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
mangum
pydantic
langchain
langchain-groq
qdrant-client
google-generativeai
langchain-google-genai